Fix undefined behavior in process_injected_command()
Running injected commands can cause inject_P() to be called with a new string.
When this happens, process_injected_command() increments injected_commands_P
incorrectly, possibly past the end of the new string, causing undefined
behavior.

To avoid this, complete adjustment of injected_commands_P prior to executing a
command. If a new call to inject_P() happens, the prior list of commands will
be interrupted and the new commands will start executing.